use 5.010. disambiguate partial commands
we have used 5.10-specific features for a while. make the dependency
explicit (and start using some of the other newfangled stuff).

we now attempt to disambiguate partial commands. this makes is possible
to use shorter commands as long as they are unique. so instead of radio
playlist you could now use radio pl. and there was much rejoice as keys
were spared.

for your changes. Lines starting